Remove node_modules from repository.
* Remove node_modules/* from repository. This means that another
  $ npm install
  will be necessary!
* Added node_modules/* to .gitignore.

I'm not sure what best practise is on this, yet suddenly npm started to
recompile certain things (bcrypt) after I added jade as a dependency. Obviously,
we need to make sure that we get the right versions of the dependencies to ensure
that the project runs properly. However, in my opinien specifying the versions
explicitly is the better approach compared to storing the deps in the repo directly.